+++ tests/testthat/test-fct_relevel.R
@@ -13,3 +13,10 @@ test_that("moves supplied levels to front", {
   f2 <- fct_relevel(f1, "c", "b")
   expect_equal(levels(f2), c("c", "b", "a", "d"))
 })
+
+test_that("can moves supplied levels to end", {
+  f1 <- factor(c("a", "b", "c", "d"))
+
+  f2 <- fct_relevel(f1, "a", "b", after = 2)
+  expect_equal(levels(f2), c("c", "d", "a", "b"))
+})